Moved balanced_accuracy into a metrics module
